implement basic cell selection
